PowerVR Graphics SDK v5.7 Release. SDK 21.1(v5.7) release.
Native SDK:
* Added support for the Khronos Vulkan Ray Tracing extension. Adding new Ray Tracing examples: Hello Ray Tracing, Hybrid Hard Shadows, Hybrid Soft Shadows, and Hybrid Reflections (currently only compatible with RT ready desktops and laptops, running windows or linux, that support Vulkan version 1.2.162.1).
* Updated Vulkan version to 1.2.162.1, glslang version to 11.2 and included the glslang validator binaries in the PowerVR Native SDK repository.

PVRVFrame:
* Fixed the pixel buffer size check for pixel operations on packed formats.
* glBindSampler now returns GL_INVALID_VALUE if the texture unit index is greater than the current profile’s GL_MAX_COMBINED_TEXTURE_IMAGE_UNITS value.
* GL_OES_vertex_array_object is now only added to the extension string if it can be emulated.
* Fixed a bug where calling glGetFramebufferAttachmentParameteriv on an attachment created with glFramebufferTexture2DMultisample* could return an invalid enum.

Refer to our release notes for more details about the changes:
https://developer.imaginationtech.com/tools/release-notes/2021-r1/

+++ framework/PVRVk/DescriptorSetVk.cpp
@@ -31,9 +31,9 @@ DescriptorPool_::DescriptorPool_(make_shared_enabler, const DeviceWeakPtr& devic
 	descPoolInfo.pNext = NULL;
 	descPoolInfo.maxSets = _createInfo.getMaxDescriptorSets();
 	descPoolInfo.flags = static_cast<VkDescriptorPoolCreateFlags>(DescriptorPoolCreateFlags::e_FREE_DESCRIPTOR_SET_BIT);
-	VkDescriptorPoolSize poolSizes[static_cast<uint32_t>(descriptorTypeSize)];
+	VkDescriptorPoolSize poolSizes[descriptorTypeSize];
 	uint32_t poolIndex = 0;
-	for (uint32_t i = 0; i < static_cast<uint32_t>(descriptorTypeSize); ++i)
+	for (uint32_t i = 0; i < descriptorTypeSize; ++i)
 	{
 		uint32_t count = _createInfo.getNumDescriptorTypes(DescriptorType(i));
 		if (count)
